+++ packages/integration-sdk/CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to
 
 ## Unreleased
 
+### Changed
+
+- Introduced `shouldReportErrorToOperator` function for integration runtime
+  environments to check whether an `Error` is of a type that need not be alerted
+  to deployment operators. All `Error`s other than those only an end user can
+  resolve should be reported.
+
 ## 3.2.0 - 2020-09-01
 
 ### Changed