+++ packages/integration-sdk/CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to
 
 ### Added
 
+- Added `dotenv/config` to exported `dev-tools` jest config.
+
+## 3.3.0 - 2020-09-25
+
+### Added
+
 - `j1-integration visualize-type` command generates metadata graph of types
 - Added `--disable-schema-validation` flag to `j1-integration collect` command.
 - Added optional `setData` input to `createMockJobState()` test util.